% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseFinding-package.R
\docType{data}
\name{biom}
\alias{biom}
\title{Biometrics Dose Response data}
\format{
A data frame with 100 observations on the following 2 variables.
\describe{
\item{\code{resp}}{a numeric vector containing the response values}
\item{\code{dose}}{a numeric vector containing the dose values}
}
}
\source{
Bretz, F., Pinheiro, J. C., and Branson, M. (2005), Combining
multiple comparisons and modeling techniques in dose-response studies,
\emph{Biometrics}, \strong{61}, 738--748
}
\usage{
data(biom)
}
\description{
An example data set for dose response studies. This data set was used in
Bretz et al. (2005) to illustrate the MCPMod methodology.
}
\keyword{datasets}
