\name{DoseFinding-internal}
\title{DoseFinding package internal functions}
\alias{getDose}
\alias{getData}
\alias{getGrid}
\alias{getInitBeta}
\alias{getInitLRasymp}
\alias{getInitP}
\alias{getLinCoef}
\alias{getOptCont}
\alias{getStdCont}
\alias{getTstat.cov}
\alias{getTstat.nocov}
\alias{getZmat}
\alias{getZmat.weighted}
\alias{modContr}
\alias{optGrid}
\alias{optLoc}
\alias{checkModels}
\alias{pValues}
\alias{sepCoef}
\alias{intervals}
\alias{fitModel.bndnls}
\alias{fitModel.lin}
\alias{fitModel.nls}
\alias{fitModel.userMod}
\alias{getTstat}
\alias{getDirection}
\alias{print.doseEst}
\alias{print.DRMod}
\alias{print.MCPtest}

\description{
  These functions are not meant to be called by the user.
}

\keyword{ internal }
