% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDFdata.R
\name{genDFdata}
\alias{genDFdata}
\alias{genDFdataNorm}
\alias{genDFdataBin}
\alias{genDFdataCount}
\title{Simulate dose-response data}
\usage{
genDFdataNorm(model, argsMod, doses, n, sigma, mu = NULL)

genDFdataBin(
  model,
  argsMod,
  doses,
  n,
  mu = NULL,
  scale = c("logit", "identity", "probit", "cloglog")
)

genDFdataCount(
  model,
  argsMod,
  doses,
  n,
  mu = NULL,
  theta = NULL,
  scale = c("log", "identity")
)
}
\arguments{
\item{model}{Character string giving the name of a model function.}

\item{argsMod}{vector with arguments for the model function}

\item{doses}{vector of dose levels to be used}

\item{n}{vector of group sample sizes}

\item{sigma}{standard deviation. Can be a single value or a vector to simulate heteroscedasticity across doses}

\item{mu}{mean vector to use if no model is specified}

\item{scale}{specifies scale on which the dose-response model or mean vector is specified. The defaults are logit scale for
binary data and log scale for count data.}

\item{theta}{parameter controlling the overdispersion of the negative binomial (see \code{\link[MASS:rnegbin]{MASS::rnegbin()}}). If not
specified poisson distribution will be used.}
}
\value{
A data frame with two columns corresponding to dose and simulated response
}
\description{
Functions to generate a dataset with simulated normally distributed, binary or count responses, according to a pre-specified
dose-response model (or mean vector)
}
\examples{
# Simulate normally distributed data under Emax model
genDFdataNorm("emax", c(e0 = 0.2, eMax = 1, ed50 = 0.05), 
              doses = c(0,0.05,0.2,0.6,1), n = 20, sigma = 1)
# use fixed mean vector
genDFdataNorm(mu = 1:5, doses = 0:4, n = c(20, 20, 10, 5, 1), sigma = 0.2)
# simulate heteroscedastic data
genDFdataNorm("emax", c(e0 = 0.2, eMax = 1, ed50 = 0.05), 
              doses = c(0,0.05,0.2,0.6,1), n = 20, sigma = seq(0.6, 1, 0.1))

# Simulate binary data under Emax model
genDFdataBin("emax", c(e0 = log(0.4/0.6), eMax = log(0.6/0.4), ed50 = 0.05), 
             doses = c(0,0.05,0.2,0.6,1), n = 20)
# use fixed mean vector on identity scale
genDFdataBin(mu = c(0.2, 0.25, 0.4, 0.5, 0.6), 
             doses = 0:4, n = c(20, 20, 10, 5, 1), scale = "identity")

# Simulate count data under Emax model
genDFdataCount("emax", c(e0 = log(0.4), eMax = log(0.6), ed50 = 0.05), 
               doses = c(0,0.05,0.2,0.6,1), n = 20)
# negative binomially distributed responses
genDFdataCount("emax", c(e0 = log(0.4), eMax = log(0.6), ed50 = 0.05), 
               doses = c(0,0.05,0.2,0.6,1), n = 20, theta = 0.2)
# use fixed mean vector on identify scale
genDFdataCount(mu = c(0.2, 0.25, 0.4, 0.5, 0.6), 
               doses = 0:4, n = c(20, 20, 10, 5, 1), scale = "identity")
}
