% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseFinding-package.R
\docType{data}
\name{IBScovars}
\alias{IBScovars}
\title{Irritable Bowel Syndrome Dose Response data with covariates}
\format{
A data frame with 369 observations on the following 2 variables.
\describe{
\item{\code{gender}}{a factor specifying the gender}
\item{\code{dose}}{a numeric vector}
\item{\code{resp}}{a numeric vector}
}
}
\source{
\insertRef{biesheuvel2002}{DoseFinding}
}
\usage{
data(IBScovars)
}
\description{
A subset of the data used by (Biesheuvel and Hothorn, 2002).  The data are
part of a dose ranging trial on a compound for the treatment of the
irritable bowel syndrome with four active treatment arms, corresponding to
doses 1,2,3,4 and placebo. Note that the original dose levels have been
blinded in this data set for confidentiality. The primary endpoint was a
baseline adjusted abdominal pain score with larger values corresponding to a
better treatment effect. In total 369 patients completed the study, with
nearly balanced allocation across the doses.
}
\keyword{datasets}
