\name{logLik.DRMod}
\Rdversion{1.1}
\alias{logLik.DRMod}
\title{
  Calculate log-likelihood value for a DRMod object
}
\description{
  Calculate log-likelihood value for a DRMod object
}
\usage{
\method{logLik}{DRMod}(object, ...)
}
\arguments{
  \item{object}{
    A DRMod object
  }
  \item{\dots}{
    Additional parameters
  }
}

\seealso{
\code{\link{fitDRModel}}, \code{\link{AIC.DRMod}}
}
\keyword{ misc }

