% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powMCTInterim.R
\name{powMCTInterim}
\alias{powMCTInterim}
\title{Calculate conditional or predictive power for multiple contrast test}
\usage{
powMCTInterim(
  contMat,
  mu0t,
  S0t,
  S_end,
  alpha = 0.025,
  type = c("predictive", "conditional"),
  mu_assumed = NULL,
  alternative = c("one.sided", "two.sided"),
  control = mvtnorm.control()
)
}
\arguments{
\item{contMat}{Contrast matrix to use. The individual contrasts should be
saved in the columns of the matrix}

\item{alpha}{Significance level to use}

\item{alternative}{Character determining the alternative for the multiple
contrast trend test.}

\item{control}{A list specifying additional control parameters for the
\samp{qmvt} and \samp{pmvt} calls in the code, see also
\samp{mvtnorm.control} for details.}

\item{altModels}{An object of class \samp{Mods}, defining the mean vectors
under which the power should be calculated}

\item{n, sigma, S}{Either a vector \samp{n} and \samp{sigma} or \samp{S} need
to be specified.  When \samp{n} and \samp{sigma} are specified it is assumed
computations are made for a normal homoscedastic ANOVA model with group
sample sizes given by \samp{n} and residual standard deviation \samp{sigma},
i.e. the covariance matrix used for the estimates is thus
\code{sigma^2*diag(1/n)} and the degrees of freedom are calculated as
\code{sum(n)-nrow(contMat)}. When a single number is specified for \samp{n}
it is assumed this is the sample size per group and balanced allocations are
used.\cr

When \samp{S} is specified this will be used as covariance matrix for the
estimates.}

\item{placAdj}{Logical, if true, it is assumed that the standard deviation
or variance matrix of the placebo-adjusted estimates are specified in
\samp{sigma} or \samp{S}, respectively. The contrast matrix has to be
produced on placebo-adjusted scale, see \code{\link{optContr}}, so that the
coefficients are no longer contrasts (i.e. do not sum to 0).}

\item{df}{Degrees of freedom to assume in case \samp{S} (a general
covariance matrix) is specified. When \samp{n} and \samp{sigma} are
specified the ones from the corresponding ANOVA model are calculated.}

\item{critV}{Critical value, if equal to \samp{TRUE} the critical value will
be calculated. Otherwise one can directly specify the critical value here.}
}
\value{
Numeric containing the calculated power values
}
\description{
Calculate predictive or conditional power for a multiple contrast test based on interim data, e.g. for a futility
interim analysis. This function can also be applied to longitudinal endpoints, where at the time of interim analysis
incomplete data is available.
}
\examples{

doses <- c(0, 0.5, 1, 2, 4, 8)
mods <- Mods(emax = c(0.5,1,2,4), sigEmax = rbind(c(0.5, 3), c(1, 3), c(2, 3), c(4,3)), quadratic = -0.1, doses = doses)
w <- c(1,0.5,0.5,0.5,1,1)
contMat <- optContr(models=mods,w=w)$contMat
sigma <- 0.3
n_final <- round(531*w/sum(w))
n <- floor(n_final/2)
S0t <- diag(sigma^2/n)
S_end <- diag(sigma^2/n_final)
## assumed interim estimate
mu0t <- 0.05*doses/(doses+1) + rnorm(6,0,0.382/sqrt(n))
## assumed mu (needed for conditional power)
mu_assumed <- 0.135*doses/(doses+1)
## compare simulation based and numerical integration approach
powMCTInterim(contMat = contMat, S0t = S0t, S_end = S_end, mu0t = mu0t, type = 'predictive')
powMCTInterim(contMat = contMat, S0t = S0t, S_end = S_end, mu0t = mu0t, type = 'conditional', mu_assumed = mu_assumed)
}
\references{
Bornkamp, B., Zhou, J., Xi, D. and Cao W. (2025). Futility analyses for the MCP-Mod methodology based
on longitudinal models, \emph{arXiv:2406.19965}
}
\seealso{
\code{\link{powMCT}} \code{\link{MCTtest}}, \code{\link{optContr}}
}
