% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoseFinding-package.R
\docType{package}
\name{DoseFinding-package}
\alias{DoseFinding}
\alias{DoseFinding-package}
\title{DoseFinding: Planning and Analyzing Dose Finding Experiments}
\description{
The DoseFinding package provides functions for the design and analysis of dose-finding experiments (for example pharmaceutical Phase II clinical trials). It provides functions for: multiple
contrast tests (MCTtest), fitting non-linear dose-response models (fitMod), a combination of testing and dose-response modelling (MCPMod), and calculating optimal designs (optDesign), both for
normal and general response variable.
}
\details{
The main functions are:\cr
\strong{MCTtest}: Implements a multiple contrast tests\cr
\strong{powMCT}: Power calculations for multiple contrast tests\cr
\strong{fitMod}: Fits non-linear dose-response models\cr
\strong{optDesign}: Calculates optimal designs for dose-response models\cr
\strong{MCPMod}: Performs MCPMod methodology\cr
\strong{sampSize}: General function for sample size calculation\cr
}
\examples{

data(IBScovars)

## perform (model based) multiple contrast test
## define candidate dose-response shapes
models <- Mods(linear = NULL, emax = 0.2, quadratic = -0.17,
               doses = c(0, 1, 2, 3, 4))
## plot models
plot(models)
## perform multiple contrast test
test <- MCTtest(dose, resp, IBScovars, models=models,
                addCovars = ~ gender)

## fit non-linear emax dose-response model
fitemax <- fitMod(dose, resp, data=IBScovars, model="emax",
                  bnds = c(0.01,5))
## display fitted dose-effect curve
plot(fitemax, CI=TRUE, plotData="meansCI")

## Calculate optimal designs for target dose (TD) estimation
doses <- c(0, 10, 25, 50, 100, 150)
fmodels <- Mods(linear = NULL, emax = 25, exponential = 85,
                logistic = c(50, 10.8811),
                doses = doses, placEff=0, maxEff=0.4)
plot(fmodels, plotTD = TRUE, Delta = 0.2)
weights <- rep(1/4, 4)
desTD <- optDesign(fmodels, weights, Delta=0.2, designCrit="TD")

}
\references{
\insertRef{bornkamp2011}{DoseFinding}

\insertRef{bornkamp2009}{DoseFinding}

\insertRef{bretz2005}{DoseFinding}

\insertRef{dette2008}{DoseFinding}

\insertRef{oquigley2017}{DoseFinding}

\insertRef{pinheiro2006}{DoseFinding}

\insertRef{pinheiro2014}{DoseFinding}

\insertRef{seber2003}{DoseFinding}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/DoseFinding}
  \item \url{https://openpharma.github.io/DoseFinding/}
  \item Report bugs at \url{https://github.com/openpharma/DoseFinding/issues}
}

}
\author{
\strong{Maintainer}: Marius Thomas \email{marius.thomas@novartis.com}

Authors:
\itemize{
  \item Bjoern Bornkamp \email{bjoern.bornkamp@novartis.com} (\href{https://orcid.org/0000-0002-6294-8185}{ORCID})
  \item Jose Pinheiro
  \item Frank Bretz
  \item Ludger Sandig
  \item Daniel Sabanes Bove \email{daniel.sabanes_bove@rconis.com} (\href{https://orcid.org/0000-0002-0176-9239}{ORCID})
}

Other contributors:
\itemize{
  \item Novartis Pharma AG [copyright holder, funder]
}

}
